Added pad_plus.c for oled. Added config.h for pad_plus. Applied Suggested Changes.
