+++ keyboards/togkey/pad_plus/config.h
@@ -1,0 +1,13 @@
+/* Copyright 2025 TogKey */
+/* SPDX-License-Identifier: GPL-2.0-or-later */
+
+#pragma once
+
+//OLED
+#define OLED_DISPLAY_128x64
+#define I2C_DRIVER I2CD1
+#define I2C1_SDA_PIN GP2
+#define I2C1_SCL_PIN GP3
+
+//Encoder
+#define ENCODER_MAP_KEY_DELAY 10
