+++ keyboards/togkey/pad_plus/keymaps/default/keymap.c
@@ -5,9 +5,9 @@
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
     [0] = LAYOUT_2x3_knob(
-                                                QK_KB_0,   
-        KC_AUDIO_VOL_DOWN,   KC_AUDIO_MUTE, KC_AUDIO_VOL_UP,
-        KC_MEDIA_PREV_TRACK, KC_MEDIA_PLAY_PAUSE, KC_MEDIA_NEXT_TRACK
+        QK_KB_0,   
+        KC_VOLD, KC_MUTE, KC_VOLU,
+        KC_MPRV, KC_MPLY, KC_MNXT
     ),
 
     [1] = LAYOUT_2x3_knob(
@@ -28,84 +28,3 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
         KC_M,   KC_N, KC_O
     )
 };
-
-#if defined(ENCODER_MAP_ENABLE)
-const uint16_t PROGMEM encoder_map[][NUM_ENCODERS][2] = {[0] = {ENCODER_CCW_CW(KC_AUDIO_VOL_DOWN, KC_AUDIO_VOL_UP)}};
-#endif
-
-#if defined(OLED_ENABLE)
-oled_rotation_t oled_init_user(oled_rotation_t rotation) {
-    return OLED_ROTATION_180;
-}
-
-enum custom_keycodes { LAYER_SCROLL = QK_KB_0 };
-
-enum blender_keycode { TOG_LAY = QK_KB_0 };
-
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-    switch (keycode) {
-        case LAYER_SCROLL:
-            if (!record->event.pressed) {
-                // Already handled the keycode, no further processing is needed
-                return false;
-            }
-
-            uint8_t current_layer = get_highest_layer(layer_state);
-            if (current_layer > 3 || current_layer < 0) {
-                return false;
-            }
-
-            uint8_t next_layer = current_layer + 1;
-            if (next_layer > 3) {
-                next_layer = 0;
-            }
-
-            layer_move(next_layer);
-
-            return false;
-        default:
-            return true; // Process all other keycodes normally
-    }
-}
-
-bool oled_task_user() {
-    static const char image_layer_1[] PROGMEM = {
-        0, 0, 248, 248, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 248, 240, 0, 0, 0, 248, 24,  8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 240, 0, 0, 0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 248, 0, 0, 0, 0, 248, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 251, 249, 1, 1, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 142, 199, 231, 119, 127, 62, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 134, 103, 115, 119, 255, 254, 204, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 192, 240, 252, 158, 255, 255, 255, 128, 128, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0,
-        0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 251, 251, 248, 248, 251, 251, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 255, 128, 0, 0, 0, 0, 0, 0, 0, 0, 7, 7, 7, 6, 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 3, 7, 7, 6, 6, 7, 7, 3, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 7, 7, 7, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 1,   1,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3, 3, 3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   1,   0,   0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,  1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0,   0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   0, 0, 0, 0, 1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0,
-    };
-
-    static const char image_layer_2[] PROGMEM = {
-        0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 248, 248, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 248, 240, 0, 0, 0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 248, 0, 0, 0, 0, 248, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 14, 254, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 249, 125, 189, 157, 193, 227, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 134, 103, 115, 119, 255, 254, 204, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 192, 240, 252, 158, 255, 255, 255, 128, 128, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0,
-        0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 7, 7, 7, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 251, 248, 251, 251, 251, 251, 251, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 3, 7, 7, 6, 6, 7, 7, 3, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 7, 7, 7, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,  1,  1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0, 1,   1,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   1,   0,   0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   0, 0, 0, 0, 1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0,
-    };
-
-    static const char image_layer_3[] PROGMEM = {
-        0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 248, 24,  8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 240, 0, 0, 0, 240, 248, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 248, 248, 0, 0, 0, 0, 248, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 14, 254, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 142, 199, 231, 119, 127, 62, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 251, 253, 221, 221, 137, 49, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 192, 240, 252, 158, 255, 255, 255, 128, 128, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0,
-        0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 7, 7, 7, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 128, 0, 0, 0, 0, 0, 0, 0, 0, 7, 7, 7, 6, 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 252, 249, 251, 251, 249, 252, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 7, 7, 7, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,  1,  1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,  1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0,   0, 0, 0, 0,   1,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,  3,   3,   3,   3,   3,   3,   3,   3,   3,   1,   1,   0, 0, 0, 0, 1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0,
-    };
-
-    static const char image_layer_4[] PROGMEM = {
-        0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 248, 24,  8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 240, 0, 0, 0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 248, 0, 0, 0, 0, 248, 248, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 248, 248, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 14, 254, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 142, 199, 231, 119, 127, 62, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 134, 103, 115, 119, 255, 254, 204, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 127, 31, 79, 115, 121, 1, 127, 127, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0,
-        0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 7, 7, 7, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 128, 0, 0, 0, 0, 0, 0, 0, 0, 7, 7, 7, 6, 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 3, 7, 7, 6, 6, 7, 7, 3, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 248, 254, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,  1,  1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,  1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0,   0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   0, 0, 0, 0, 1,   1,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,  3,  3,   3,   3, 3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   1,   1,   0, 0, 0, 0,
-    };
-
-    switch (get_highest_layer(layer_state)) {
-        case 0:
-            oled_write_raw_P(image_layer_1, sizeof(image_layer_1));
-            break;
-        case 1:
-            oled_write_raw_P(image_layer_2, sizeof(image_layer_2));
-            break;
-        case 2:
-            oled_write_raw_P(image_layer_3, sizeof(image_layer_3));
-            break;
-        case 3:
-            oled_write_raw_P(image_layer_4, sizeof(image_layer_4));
-            break;
-        default:
-            oled_write_ln_P(PSTR("Undefined"), false);
-    }
-
-    return false;
-}
-#endif
