+++ keyboards/togkey/pad_plus/pad_plus.c
@@ -1,0 +1,79 @@
+/* Copyright 2025 TogKey */
+/* SPDX-License-Identifier: GPL-2.0-or-later */
+
+#include "quantum.h"
+
+#if defined(OLED_ENABLE)
+oled_rotation_t oled_init_user(oled_rotation_t rotation) {
+    return OLED_ROTATION_180;
+}
+
+enum custom_keycodes { LAYER_SCROLL = QK_KB_0 };
+
+bool process_record_user(uint16_t keycode, keyrecord_t *record) {
+    switch (keycode) {
+        case LAYER_SCROLL:
+            if (!record->event.pressed) {
+                // Already handled the keycode, no further processing is needed
+                return false;
+            }
+
+            uint8_t current_layer = get_highest_layer(layer_state);
+            if (current_layer > 3 || current_layer < 0) {
+                return false;
+            }
+
+            uint8_t next_layer = current_layer + 1;
+            if (next_layer > 3) {
+                next_layer = 0;
+            }
+
+            layer_move(next_layer);
+
+            return false;
+        default:
+            return true; // Process all other keycodes normally
+    }
+}
+
+bool oled_task_user() {
+    static const char image_layer_1[] PROGMEM = {
+        0, 0, 248, 248, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 248, 240, 0, 0, 0, 248, 24,  8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 240, 0, 0, 0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 248, 0, 0, 0, 0, 248, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 251, 249, 1, 1, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 142, 199, 231, 119, 127, 62, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 134, 103, 115, 119, 255, 254, 204, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 192, 240, 252, 158, 255, 255, 255, 128, 128, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0,
+        0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 251, 251, 248, 248, 251, 251, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 255, 128, 0, 0, 0, 0, 0, 0, 0, 0, 7, 7, 7, 6, 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 3, 7, 7, 6, 6, 7, 7, 3, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 7, 7, 7, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 1,   1,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3, 3, 3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   1,   0,   0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,  1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0,   0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   0, 0, 0, 0, 1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0,
+    };
+
+    static const char image_layer_2[] PROGMEM = {
+        0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 248, 248, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 248, 240, 0, 0, 0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 248, 0, 0, 0, 0, 248, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 14, 254, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 249, 125, 189, 157, 193, 227, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 134, 103, 115, 119, 255, 254, 204, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 192, 240, 252, 158, 255, 255, 255, 128, 128, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0,
+        0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 7, 7, 7, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 251, 248, 251, 251, 251, 251, 251, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 3, 7, 7, 6, 6, 7, 7, 3, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 7, 7, 7, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,  1,  1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0, 1,   1,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   1,   0,   0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   0, 0, 0, 0, 1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0,
+    };
+
+    static const char image_layer_3[] PROGMEM = {
+        0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 248, 24,  8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 240, 0, 0, 0, 240, 248, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 248, 248, 0, 0, 0, 0, 248, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 14, 254, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 142, 199, 231, 119, 127, 62, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 251, 253, 221, 221, 137, 49, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 192, 240, 252, 158, 255, 255, 255, 128, 128, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0,
+        0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 7, 7, 7, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 128, 0, 0, 0, 0, 0, 0, 0, 0, 7, 7, 7, 6, 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 252, 249, 251, 251, 249, 252, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 7, 7, 7, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,  1,  1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,  1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0,   0, 0, 0, 0,   1,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,  3,   3,   3,   3,   3,   3,   3,   3,   3,   1,   1,   0, 0, 0, 0, 1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0,
+    };
+
+    static const char image_layer_4[] PROGMEM = {
+        0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 0, 0, 0, 0, 248, 24,  8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 240, 0, 0, 0, 0, 248, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 248, 248, 0, 0, 0, 0, 248, 248, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 248, 248, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 14, 254, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 142, 199, 231, 119, 127, 62, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 134, 103, 115, 119, 255, 254, 204, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 127, 31, 79, 115, 121, 1, 127, 127, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0,
+        0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 7, 7, 7, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 0, 0, 0, 0, 255, 128, 0, 0, 0, 0, 0, 0, 0, 0, 7, 7, 7, 6, 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 0, 0, 0, 0, 0, 0, 0, 0, 3, 7, 7, 6, 6, 7, 7, 3, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 248, 254, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,  1,  1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,  1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   0,   0, 0, 0, 0, 1,   1, 1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   1,   1,   1,   1,   1,   1, 1, 1, 1, 1, 1, 1, 1, 1,   1,   0, 0, 0, 0, 1,   1,   3,   3,   3,   3,   3,   3,   3,   3,   3,   3,  3,  3,   3,   3, 3,   3,   3,   3,   3,   3,   3,   3,   3,   3,   1,   1,   0, 0, 0, 0,
+    };
+
+    switch (get_highest_layer(layer_state)) {
+        case 0:
+            oled_write_raw_P(image_layer_1, sizeof(image_layer_1));
+            break;
+        case 1:
+            oled_write_raw_P(image_layer_2, sizeof(image_layer_2));
+            break;
+        case 2:
+            oled_write_raw_P(image_layer_3, sizeof(image_layer_3));
+            break;
+        case 3:
+            oled_write_raw_P(image_layer_4, sizeof(image_layer_4));
+            break;
+        default:
+            oled_write_ln_P(PSTR("Undefined"), false);
+    }
+
+    return false;
+}
+#endif
